pyproject.toml
[tool.poetry]
name = "dhex"
version = "0.1.0"
description = ""
authors = ["ck1e <ck1e.wk@gmail.com@>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
